+++ CHANGELOG.md
@@ -1,12 +1,23 @@
 # Changelog
 
-<!-- spell-checker: ignore markdownlint -->
+<!-- spell-checker: ignore markdownlint volterra -->
 <!-- markdownlint-disable MD024 -->
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.2.0] - 2021-04-01
+
+### Added
+
+- Service account module to create a GCP service account, with binding to custom
+  IAM role, with credentials stored in Volterra tenant as a GCP Cloud Credential.
+
+### Changed
+
+### Removed
 
 ## [0.1.0] - 2021-04-01
 
@@ -18,7 +29,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-<!-- TODO: @memes - fix links
-[1.0.1]: https://github.com/memes/terraform-google-volterra-vpc/compare/1.0.0...1.0.1
--->
+[0.2.0]: https://github.com/memes/terraform-google-volterra/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/memes/terraform-google-volterra/releases/tag/0.1.0